🎸 feat(dynenv): support authentication on dynamic environment
Spotify (OAuth2) needs the redirect urls specified in their auth configuration.
Since netlify uses dynamic deployments to provide previews on pull requests we'd need to add every possible dynamic environment target to this list.
Due to security reasons it is not allowed to provide wildcard targets.
Therefore we let the request redirect to our production site which then checks where the request originates from.
If it is in fact a dynamic environment then we'll redirect to that, otherwise the authentication continues.

Closes #8